+++ lib/src/controls/navigation/navigation_view/view.dart
@@ -1,5 +1,4 @@
 import 'dart:math';
-import 'dart:ui' as ui;
 
 import 'package:fluent_ui/fluent_ui.dart';